% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.AdaDelta.R
\name{adaDelta}
\alias{adaDelta}
\title{Optimize mathematical function using the RMSProp algorithm}
\usage{
adaDelta(f, x0, max.iter = 100, gamma = 0.95, stop.grad = .Machine$double.eps)
}
\arguments{
\item{f}{a (multi-) dimensional function to be eptimized.}

\item{x0}{the starting point of the optimization.}

\item{max.iter}{the maximum number of iterations performed in the optimization.}

\item{stop.grad}{the stop-criterion for the gradient change.}
}
\description{
This functions uses the RMSProp algorithm to find the minimum of a (multi-)
dimensional mathematical function. The algorithm searches for the.
The 'eps' factor avoids numerical issues of dividing by 0. The 'step.size'
scales the movement into the single coordinate direction.
}
