% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.2d.R
\name{plot2d}
\alias{plot2d}
\title{Visualize optimization procedure in 2d contour plot}
\usage{
plot2d(f, x1.lower, x1.upper, x2.lower, x2.upper, n.x, xmat)
}
\arguments{
\item{f}{a (multi-) dimensional function to be eptimized.}

\item{x1.lower}{the lower boundary for the range of the x1 coordinate displayed.}

\item{x1.upper}{the upper boundary for the range of the x1 coordinate displayed.}

\item{x2.lower}{the lower boundary for the range of the x2 coordinate displayed.}

\item{x2.upper}{the upper boundary for the range of the x2 coordinate displayed.}

\item{n.x}{the number (equidistant) points at which f is evaluated w.r.t x1 and x2 coordinate.}

\item{xmat}{an object of class}
}
\description{
This functions visualizes the steps of the optimization in a 2 dimensional contour plot.
}
