% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.2d.info.R
\name{plot2d_info}
\alias{plot2d_info}
\title{Visualize optimization procedure in 2d contour plot}
\usage{
plot2d_info(
  f,
  x1.lower,
  x1.upper,
  x2.lower,
  x2.upper,
  n.x = 30L,
  xmat,
  trueOpt = NULL,
  algoName = NULL,
  plotSegments = FALSE,
  optimError = FALSE
)
}
\arguments{
\item{f}{a (multi-) dimensional function to be eptimized.}

\item{x1.lower}{the lower boundary for the range of the x1 coordinate displayed.}

\item{x1.upper}{the upper boundary for the range of the x1 coordinate displayed.}

\item{x2.lower}{the lower boundary for the range of the x2 coordinate displayed.}

\item{x2.upper}{the upper boundary for the range of the x2 coordinate displayed.}

\item{n.x}{the number (equidistant) points at which f is evaluated w.r.t x1 and x2 coordinate.}

\item{xmat}{an object of class list containing the results of the algorithm.}

\item{trueOpt}{is a numeric vector of the true optimum of the function.}

\item{optimError}{passed argument from optimization. Specifies if error in gradient occured.}

\item{algoNamw}{list of strings with names of algorithms. Length must equal length of 'xmat'.}
}
\description{
This functions visualizes the steps of the optimization in a 2 dimensional contour plot.
The default upper and lower bounds for the plot are set by the box constraints of the
respective optimization function.
}
