% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loss.R
\name{plotLoss}
\alias{plotLoss}
\title{Visualize loss of optimization procedure}
\usage{
plotLoss(
  f,
  x1.lower,
  x1.upper,
  x2.lower,
  x2.upper,
  xmat,
  trueOptZ = NULL,
  algoName = NULL,
  optimError = FALSE
)
}
\arguments{
\item{f}{a (multi-) dimensional function to be eptimized.}

\item{x1.lower}{the lower boundary for the range of the x1 coordinate displayed.}

\item{x1.upper}{the upper boundary for the range of the x1 coordinate displayed.}

\item{x2.lower}{the lower boundary for the range of the x2 coordinate displayed.}

\item{x2.upper}{the upper boundary for the range of the x2 coordinate displayed.}

\item{xmat}{an object of class list containing the results of the algorithm.}

\item{trueOptZ}{is a numeric of the z coordinate of the true optimum.}

\item{algoName}{list of strings with names of algorithms. Length must equal length of 'xmat'.}

\item{optimError}{passed argument from optimization. Specifies if error in gradient occured.}

\item{n.x}{the number (equidistant) points at which f is evaluated w.r.t x1 and x2 coordinate.}
}
\description{
This functions visualizes the difference between the true optimum and the optimum in the
current iteration (loss) during the optimization procedure.
The default upper and lower bounds for the plot are set by the box constraints of the
respective optimization function.
}
