% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAppLocal.R
\name{runAppLocal}
\alias{runAppLocal}
\title{Run Shiny App on local machine}
\usage{
runAppLocal()
}
\description{
This function helps to deploy the shiny app in an easy way
on the local machine. As alternativ you can use the app
deployed on the server which can be found in the Git repo.
}
